+++ src/commands/moderation/slowmode.ts
@@ -1,14 +1,9 @@
 import { Colors, EmbedField, TextChannel, ChannelType } from 'discord.js';
 import { CommandDefinition } from '../../lib/command';
-import { Roles, Channels, CommandCategory } from '../../constants';
+import { Channels, CommandCategory, RoleGroups } from '../../constants';
 import { makeEmbed } from '../../lib/embed';
 import { getScheduler } from '../../lib/scheduler';
 import { client } from '../..';
-
-const permittedRoles = [
-    Roles.ADMIN_TEAM,
-    Roles.MODERATION_TEAM,
-];
 
 enum TimeConversions {
     SECONDS_TO_MILLISECONDS = 1000,
@@ -80,12 +75,6 @@ const noSchedulerEmbed = makeEmbed({
     color: Colors.Red,
 });
 
-const noPermEmbed = makeEmbed({
-    title: 'Slow Mode - Permission missing',
-    description: 'You do not have permission to use this command.',
-    color: Colors.Red,
-});
-
 const slowModeEmbedField = (moderator: string, channel: string, slowmode: string, timeout: string): EmbedField[] => [
     {
         inline: true,
@@ -113,16 +102,12 @@ export const slowMode: CommandDefinition = {
     name: ['slowmode'],
     description: 'Manages slow mode for channels and potentially disable slow mode after a certain time.',
     category: CommandCategory.MODERATION,
+    requirements: { roles: RoleGroups.STAFF },
     executor: async (msg) => {
         const subCommands = ['set', 'disable'];
         const scheduler = getScheduler();
         if (!scheduler) {
             await msg.channel.send({ embeds: [noSchedulerEmbed] });
-        }
-
-        const hasPermittedRole = msg.member.roles.cache.some((role) => permittedRoles.map((r) => r.toString()).includes(role.id));
-        if (!hasPermittedRole) {
-            return msg.channel.send({ embeds: [noPermEmbed] });
         }
 
         const modLogsChannel = client.channels.resolve(Channels.MOD_LOGS) as TextChannel | null;