+++ src/commands/moderation/temporaryCommandEdit.ts
@@ -1,15 +1,10 @@
 import { Colors, EmbedField, TextChannel } from 'discord.js';
 import moment from 'moment';
 import { CommandDefinition } from '../../lib/command';
-import { Roles, Channels, CommandCategory } from '../../constants';
+import { Channels, CommandCategory, RoleGroups } from '../../constants';
 import { makeEmbed, makeLines } from '../../lib/embed';
 import { getConn } from '../../lib/db';
 import TemporaryCommand from '../../lib/schemas/temporaryCommandSchema';
-
-const permittedRoles = [
-    Roles.ADMIN_TEAM,
-    Roles.MODERATION_TEAM,
-];
 
 const helpEmbed = (evokedCommand: String) => makeEmbed({
     title: 'Manage Temporary Commands - Help',
@@ -133,12 +128,6 @@ const noConnEmbed = makeEmbed({
     color: Colors.Red,
 });
 
-const noPermEmbed = makeEmbed({
-    title: 'Temporary Command - Permission missing',
-    description: 'You do not have permission to use this command.',
-    color: Colors.Red,
-});
-
 const notFoundEmbed = (action: string, command: string) => makeEmbed({
     title: `Temporary Command - ${action} - ${command} not found`,
     description: 'No Temporary Command(s) matching the search can be found.',
@@ -227,6 +216,7 @@ export const temporarycommandedit: CommandDefinition = {
     name: ['temporarycommandedit', 'tempcommandedit', 'tcedit', 'tcmod'],
     description: 'Creates a temporary command for temporary use.',
     category: CommandCategory.MODERATION,
+    requirements: { roles: RoleGroups.STAFF },
     executor: async (msg) => {
         const subCommands = ['add', 'image', 'delete', 'info'];
         const conn = await getConn();
@@ -236,12 +226,8 @@ export const temporarycommandedit: CommandDefinition = {
 
         const modLogsChannel = msg.guild.channels.resolve(Channels.MOD_LOGS) as TextChannel | null;
         const supportOpsChannel = msg.guild.channels.resolve(Channels.SUPPORT_OPS) as TextChannel | null;
-        const hasPermittedRole = msg.member.roles.cache.some((role) => permittedRoles.map((r) => r.toString()).includes(role.id));
         const evokedCommand = msg.content.split(/\s+/)[0];
         const args = msg.content.replace(evokedCommand, '').trim();
-        if (!hasPermittedRole) {
-            return msg.channel.send({ embeds: [noPermEmbed] });
-        }
         if (!args || args === 'help') {
             return msg.channel.send({ embeds: [helpEmbed(evokedCommand)] });
         }