+++ src/commands/utils/birthday.ts
@@ -1,23 +1,19 @@
 import { Colors } from 'discord.js';
 import { CommandDefinition } from '../../lib/command';
-import { Roles, CommandCategory, Threads } from '../../constants';
+import { RoleGroups, CommandCategory, Threads } from '../../constants';
 import { makeEmbed, makeLines } from '../../lib/embed';
 import Logger from '../../lib/logger';
 import Birthday from '../../lib/schemas/birthdaySchema';
 import { getConn } from '../../lib/db';
 
-const permittedRoles = [
-    Roles.ADMIN_TEAM,
-    Roles.MODERATION_TEAM,
-    Roles.DEVELOPMENT_TEAM,
-    Roles.MEDIA_TEAM,
-    Roles.FBW_EMERITUS,
-];
-
 export const birthday: CommandDefinition = {
     name: 'birthday',
     description: 'Manages birthday reminders',
     category: CommandCategory.UTILS,
+    requirements: {
+        roles: RoleGroups.TEAM,
+        channels: [Threads.BIRTHDAY_THREAD],
+    },
     executor: async (msg) => {
         const conn = await getConn();
 
@@ -35,21 +31,7 @@ export const birthday: CommandDefinition = {
 
         let birthdayEmbed;
 
-        const hasPermittedRole = msg.member.roles.cache.some((role) => permittedRoles.map((r) => r.toString()).includes(role.id));
-
-        if (!hasPermittedRole) {
-            birthdayEmbed = makeEmbed({
-                title: 'Birthday reminder',
-                description: 'You do not have permission to use this command.',
-                color: Colors.Red,
-            });
-        } else if (msg.channel.id !== Threads.BIRTHDAY_THREAD) {
-            birthdayEmbed = makeEmbed({
-                title: 'Birthday reminder',
-                description: `That command can only be used in <#${Threads.BIRTHDAY_THREAD}>`,
-                color: Colors.Red,
-            });
-        } else if (args[0] === 'add' || args[0] === 'set') {
+        if (args[0] === 'add' || args[0] === 'set') {
             const member = msg.mentions.members.first();
 
             if (!member) {