+++ src/lib/typeCommand.ts
@@ -1,5 +1,5 @@
 import { EmbedData } from 'discord.js';
-import { CommandDefinition, isMessageCommand, MessageCommandDefinition } from './command';
+import { CommandDefinition, isMessageCommand, MessageCommandDefinition, hasRequiredPermissions, sendPermissionsEmbed } from './command';
 import { AircraftTypeList, CommandCategory } from '../constants';
 import commands from '../commands/index';
 import Logger from './logger';
@@ -37,7 +37,7 @@ export const typeCommand: CommandDefinition = {
             // Evoked with a non-existing command
             return;
         }
-        const { name } = command;
+        const { name, requirements } = command;
         const commandsArray = Array.isArray(name) ? name : [name];
         if (!commandsArray.includes(subCommand)) {
             Logger.debug(`Type Command - evoked: ${evokedCommand} - sub: ${subCommand} - Mismatch between sub command and command`);
@@ -49,78 +49,87 @@ export const typeCommand: CommandDefinition = {
             // This isn't a simple MessageCommand
             return;
         }
-        const { genericEmbed, typeEmbeds } = (command as MessageCommandDefinition);
-        if (!enableMultipleAircraftTypes || !typeEmbeds || Object.keys(typeEmbeds).length === 0) {
-            if (subCommand === evokedCommand) {
-                const postEmbed = typeEmbeds && Object.keys(typeEmbeds).includes(defaultDisabledAircraftType) ? typeEmbeds[defaultDisabledAircraftType] : genericEmbed;
-                await msg.channel.send({ embeds: [postEmbed] });
-            }
-            return;
-        }
-        const commandSupportedAircraftTypes = typeEmbeds ? Object.keys(typeEmbeds) : [];
-        if (subCommand === evokedCommand) {
-            const commandSupportedAircraftTypeEmojis = [];
-            const { data: genericEmbedData } = genericEmbed;
-            const postGenericEmbed = makeEmbed((genericEmbedData as EmbedData));
-            const choiceEmbedFieldLines = [];
-            choiceEmbedFieldLines.push('Please select the appropriate reaction for the Aircraft you would like more information about:');
-            commandSupportedAircraftTypes.forEach((typeName) => {
-                if (typeName in AircraftTypeList) {
-                    commandSupportedAircraftTypeEmojis.push(AircraftTypeList[typeName]);
-                    choiceEmbedFieldLines.push(`<:${AircraftTypeList[typeName]}> - ${typeName.toUpperCase()}`);
+
+        const member = await msg.guild.members.fetch(msg.author);
+        const [requirementsSatisfied, requirementsError] = hasRequiredPermissions(requirements, member, msg.channel.id);
+
+        if (requirementsSatisfied) {
+            const { genericEmbed, typeEmbeds } = (command as MessageCommandDefinition);
+            if (!enableMultipleAircraftTypes || !typeEmbeds || Object.keys(typeEmbeds).length === 0) {
+                if (subCommand === evokedCommand) {
+                    const postEmbed = typeEmbeds && Object.keys(typeEmbeds).includes(defaultDisabledAircraftType) ? typeEmbeds[defaultDisabledAircraftType] : genericEmbed;
+                    await msg.channel.send({ embeds: [postEmbed] });
                 }
-            });
-            postGenericEmbed.addFields({
-                name: 'Select the aircraft for more details:',
-                value: makeLines(choiceEmbedFieldLines),
-            });
-            await msg.channel.send({ embeds: [postGenericEmbed] }).then(async (genericMessage) => {
-                commandSupportedAircraftTypeEmojis.forEach(async (element) => {
-                    try {
-                        await genericMessage.react(element);
-                    } catch (e) {
-                        Logger.debug(`Failed to add reaction: ${e}`);
+                return;
+            }
+            const commandSupportedAircraftTypes = typeEmbeds ? Object.keys(typeEmbeds) : [];
+            if (subCommand === evokedCommand) {
+                const commandSupportedAircraftTypeEmojis = [];
+                const { data: genericEmbedData } = genericEmbed;
+                const postGenericEmbed = makeEmbed((genericEmbedData as EmbedData));
+                const choiceEmbedFieldLines = [];
+                choiceEmbedFieldLines.push('Please select the appropriate reaction for the Aircraft you would like more information about:');
+                commandSupportedAircraftTypes.forEach((typeName) => {
+                    if (typeName in AircraftTypeList) {
+                        commandSupportedAircraftTypeEmojis.push(AircraftTypeList[typeName]);
+                        choiceEmbedFieldLines.push(`<:${AircraftTypeList[typeName]}> - ${typeName.toUpperCase()}`);
                     }
                 });
-                const filter = (reaction, user) => commandSupportedAircraftTypeEmojis.includes(reaction.emoji.identifier) && user.id === author.id;
-                await genericMessage.awaitReactions({
-                    filter,
-                    max: 1,
-                    time: REACTION_WAIT_TIME,
-                    errors: ['time'],
-                }).then(async (collected) => {
-                    const reaction = collected.first();
-                    let foundAircraftType;
-                    for (const element of commandSupportedAircraftTypes) {
-                        if (reaction.emoji.identifier === AircraftTypeList[element]) {
-                            foundAircraftType = element;
-                            break;
-                        }
-                    }
-                    if (foundAircraftType) {
+                postGenericEmbed.addFields({
+                    name: 'Select the aircraft for more details:',
+                    value: makeLines(choiceEmbedFieldLines),
+                });
+                await msg.channel.send({ embeds: [postGenericEmbed] }).then(async (genericMessage) => {
+                    commandSupportedAircraftTypeEmojis.forEach(async (element) => {
                         try {
+                            await genericMessage.react(element);
+                        } catch (e) {
+                            Logger.debug(`Failed to add reaction: ${e}`);
+                        }
+                    });
+                    const filter = (reaction, user) => commandSupportedAircraftTypeEmojis.includes(reaction.emoji.identifier) && user.id === author.id;
+                    await genericMessage.awaitReactions({
+                        filter,
+                        max: 1,
+                        time: REACTION_WAIT_TIME,
+                        errors: ['time'],
+                    }).then(async (collected) => {
+                        const reaction = collected.first();
+                        let foundAircraftType;
+                        for (const element of commandSupportedAircraftTypes) {
+                            if (reaction.emoji.identifier === AircraftTypeList[element]) {
+                                foundAircraftType = element;
+                                break;
+                            }
+                        }
+                        if (foundAircraftType) {
+                            try {
+                                await genericMessage.delete();
+                            } catch (e) {
+                                Logger.debug('Type Command - Generic message with choices was already deleted');
+                            }
+                            await msg.channel.send({ embeds: [typeEmbeds[foundAircraftType]] });
+                        }
+                    }).catch(async () => {
+                        try {
+                            Logger.debug(`Type Command - Waited longer than ${REACTION_WAIT_TIME}ms, skipping`);
                             await genericMessage.delete();
                         } catch (e) {
                             Logger.debug('Type Command - Generic message with choices was already deleted');
                         }
-                        await msg.channel.send({ embeds: [typeEmbeds[foundAircraftType]] });
-                    }
-                }).catch(async () => {
-                    try {
-                        Logger.debug(`Type Command - Waited longer than ${REACTION_WAIT_TIME}ms, skipping`);
-                        await genericMessage.delete();
-                    } catch (e) {
-                        Logger.debug('Type Command - Generic message with choices was already deleted');
-                    }
+                    });
                 });
-            });
-            return;
+                return;
+            }
+            if (commandSupportedAircraftTypes.includes(evokedCommand)) {
+                const keyTyped = evokedCommand as keyof typeof typeEmbeds;
+                await msg.channel.send({ embeds: [typeEmbeds[keyTyped]] });
+                return;
+            }
+            Logger.debug(`Type Command - evoked: ${evokedCommand} - sub: ${subCommand} - No Embed for the specified type (evoked command)`);
+        } else {
+            Logger.debug('Type Command - Bailing due to unsatisfied command requirements');
+            await sendPermissionsEmbed(msg, requirementsError);
         }
-        if (commandSupportedAircraftTypes.includes(evokedCommand)) {
-            const keyTyped = evokedCommand as keyof typeof typeEmbeds;
-            await msg.channel.send({ embeds: [typeEmbeds[keyTyped]] });
-            return;
-        }
-        Logger.debug(`Type Command - evoked: ${evokedCommand} - sub: ${subCommand} - No Embed for the specified type (evoked command)`);
     },
 };